requirements.txt
jinjasql==0.1.7
pandas==0.24.2
pysftp==0.2.8
tensorflow==1.13.1
numpy==1.16.0
SQLAlchemy==1.3.2
requests==2.19.1
Pillow==7.0.0
